SN-48 SN-25 moved group member invite/request & approve/decline functionalities to separate handler and repository to have better readability; added targetUserId to groupInvitations table but currently is not quite implemented in the repository level; groupMemberHandler needs refactor as some of the functionalities are separated into different handlers, need to implement auth checking to the groupMemberHandler.go
