+++ backend/pkg/repository/groupRepository.go
@@ -101,94 +101,97 @@ func (r *GroupRepository) DeleteGroup(id int) error {
 	return err
 }
 
-// AddMemberToGroup adds a member to a group in the database.
-// It returns an error if any.
-func (r *GroupRepository) AddMemberToGroup(groupId, userId int) error {
-	query := `INSERT INTO group_members (group_id, user_id) VALUES (?, ?)`
-	_, err := r.db.Exec(query, groupId, userId)
-	return err
-}
+// DEPRECATED: moved to groupMemberRepository.go
+// // AddMemberToGroup adds a member to a group in the database.
+// // It returns an error if any.
+// func (r *GroupRepository) AddMemberToGroup(groupId, userId int) error {
+// 	query := `INSERT INTO group_members (group_id, user_id) VALUES (?, ?)`
+// 	_, err := r.db.Exec(query, groupId, userId)
+// 	return err
+// }
 
-// RemoveMemberFromGroup removes a member from a group in the database.
-// It returns an error if any.
-func (r *GroupRepository) RemoveMemberFromGroup(groupId, userId int) error {
-	query := `DELETE FROM group_members WHERE group_id = ? AND user_id = ?`
-	_, err := r.db.Exec(query, groupId, userId)
-	return err
-}
+// // RemoveMemberFromGroup removes a member from a group in the database.
+// // It returns an error if any.
+// func (r *GroupRepository) RemoveMemberFromGroup(groupId, userId int) error {
+// 	query := `DELETE FROM group_members WHERE group_id = ? AND user_id = ?`
+// 	_, err := r.db.Exec(query, groupId, userId)
+// 	return err
+// }
 
-// CreateGroupInvitation creates a new invitation in the database.
-// It returns an error if any.
-func (r *InvitationRepository) CreateGroupInvitation(invitation model.GroupInvitation) error {
-	query := `INSERT INTO group_invitations (group_id, user_id, status) VALUES (?, ?, ?)`
-	_, err := r.db.Exec(query, invitation.GroupId, invitation.UserId, "pending")
-	return err
-}
+// // CreateGroupInvitation creates a new invitation in the database.
+// // It returns an error if any.
+// // TODO: implement notifications for group invitations
+// func (r *InvitationRepository) CreateGroupInvitation(invitation model.GroupInvitation) error {
+// 	query := `INSERT INTO group_invitations (group_id, user_id, status) VALUES (?, ?, ?)`
+// 	_, err := r.db.Exec(query, invitation.GroupId, invitation.UserId, "pending")
+// 	return err
+// }
 
-// UpdateGroupInvitation updates the status of an invitation in the database.
-// It returns an error if any.
-func (r *InvitationRepository) UpdateGroupInvitation(invitation model.GroupInvitation) error {
-	query := `UPDATE group_invitations SET status = ? WHERE id = ?`
-	_, err := r.db.Exec(query, invitation.Status, invitation.Id)
-	return err
-}
+// // UpdateGroupInvitation updates the status of an invitation in the database.
+// // It returns an error if any.
+// // DEPRECATED: use AcceptGroupInvitation and DeclineGroupInvitation instead
+// func (r *InvitationRepository) UpdateGroupInvitation(invitation model.GroupInvitation) error {
+// 	query := `UPDATE group_invitations SET status = ? WHERE id = ?`
+// 	_, err := r.db.Exec(query, invitation.Status, invitation.Id)
+// 	return err
+// }
 
-// DeleteGroupInvitation deletes an invitation from the database.
-// It returns an error if any.
-func (r *InvitationRepository) DeleteGroupInvitation(id int) error {
-	query := `DELETE FROM group_invitations WHERE id = ?`
-	_, err := r.db.Exec(query, id)
-	return err
-}
+// // DeleteGroupInvitation deletes an invitation from the database.
+// // It returns an error if any.
+// func (r *InvitationRepository) DeleteGroupInvitation(id int) error {
+// 	query := `DELETE FROM group_invitations WHERE id = ?`
+// 	_, err := r.db.Exec(query, id)
+// 	return err
+// }
 
-// GetAllGroupInvitations retrieves all group invitations from the database.
-// It returns a slice of GroupInvitation objects and an error if any.
-func (r *InvitationRepository) GetAllGroupInvitations() ([]model.GroupInvitation, error) {
-	// SQL query to select all group invitations
-	query := `SELECT * FROM group_invitations`
-	rows, err := r.db.Query(query)
-	if err != nil {
-		return nil, err
-	}
-	defer rows.Close()
+// // GetAllGroupInvitations retrieves all group invitations from the database.
+// // It returns a slice of GroupInvitation objects and an error if any.
+// func (r *InvitationRepository) GetAllGroupInvitations() ([]model.GroupInvitation, error) {
+// 	// SQL query to select all group invitations
+// 	query := `SELECT * FROM group_invitations`
+// 	rows, err := r.db.Query(query)
+// 	if err != nil {
+// 		return nil, err
+// 	}
+// 	defer rows.Close()
 
-	var invitations []model.GroupInvitation
-	for rows.Next() {
-		var invitation model.GroupInvitation
-		if err := rows.Scan(&invitation.Id, &invitation.GroupId, &invitation.UserId, &invitation.Status); err != nil {
-			return nil, err
-		}
-		invitations = append(invitations, invitation)
-	}
+// 	var invitations []model.GroupInvitation
+// 	for rows.Next() {
+// 		var invitation model.GroupInvitation
+// 		if err := rows.Scan(&invitation.Id, &invitation.GroupId, &invitation.UserId, &invitation.Status); err != nil {
+// 			return nil, err
+// 		}
+// 		invitations = append(invitations, invitation)
+// 	}
 
-	return invitations, nil
-}
+// 	return invitations, nil
+// }
 
-// GetGroupInvitationByID retrieves an invitation by ID from the database.
-// It returns the GroupInvitation object and an error if any.
-func (r *InvitationRepository) GetGroupInvitationByID(id string) (model.GroupInvitation, error) {
-	query := `SELECT * FROM group_invitations WHERE id = ?`
-	row := r.db.QueryRow(query, id)
+// // GetGroupInvitationByID retrieves an invitation by ID from the database.
+// // It returns the GroupInvitation object and an error if any.
+// func (r *InvitationRepository) GetGroupInvitationByID(id string) (model.GroupInvitation, error) {
+// 	query := `SELECT * FROM group_invitations WHERE id = ?`
+// 	row := r.db.QueryRow(query, id)
 
-	var invitation model.GroupInvitation
-	if err := row.Scan(&invitation.Id, &invitation.GroupId, &invitation.UserId, &invitation.Status); err != nil {
-		return model.GroupInvitation{}, err
-	}
-	return invitation, nil
-}
+// 	var invitation model.GroupInvitation
+// 	if err := row.Scan(&invitation.Id, &invitation.GroupId, &invitation.UserId, &invitation.Status); err != nil {
+// 		return model.GroupInvitation{}, err
+// 	}
+// 	return invitation, nil
+// }
 
-// AcceptGroupInvitation updates the status of an invitation to "accepted" in the database.
-// It returns an error if any.
-func (r *InvitationRepository) AcceptGroupInvitation(id string) error {
-	query := `UPDATE group_invitations SET status = 'accepted' WHERE id = ?`
-	_, err := r.db.Exec(query, id)
-	return err
-}
+// // AcceptGroupInvitation updates the status of an invitation to "accepted" in the database.
+// // It returns an error if any.
+// func (r *InvitationRepository) AcceptGroupInvitation(id string) error {
+// 	query := `UPDATE group_invitations SET status = 'accepted' WHERE id = ?`
+// 	_, err := r.db.Exec(query, id)
+// 	return err
+// }
 
-// DeclineGroupInvitation updates the status of an invitation to "declined" in the database.
-// It returns an error if any.
-func (r *InvitationRepository) DeclineGroupInvitation(id string) error {
-	query := `UPDATE group_invitations SET status = 'declined' WHERE id = ?`
-	_, err := r.db.Exec(query, id)
-	return err
-}
+// // DeclineGroupInvitation updates the status of an invitation to "declined" in the database.
+// // It returns an error if any.
+// func (r *InvitationRepository) DeclineGroupInvitation(id string) error {
+// 	query := `UPDATE group_invitations SET status = 'declined' WHERE id = ?`
+// 	_, err := r.db.Exec(query, id)
+// 	return err
+// }
